requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.6.3
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
dateparser==1.1.1
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
fonttools==4.33.2
frozenlist==1.3.0
google-auth==2.6.5
google-auth-oauthlib==0.5.1
gspread==5.3.2
idna==3.3
ipykernel==6.13.0
ipython==8.2.0
jedi==0.18.1
jsonschema==4.5.1
jupyter-client==7.2.2
jupyter-core==4.10.0
kiwisolver==1.4.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
mplfinance==0.12.8b9
multidict==6.0.2
nbformat==5.4.0
nest-asyncio==1.5.5
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
pandas-ta==0.3.14b0
parso==0.8.3
Pebble==4.6.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
plotly==5.8.0
prompt-toolkit==3.0.29
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.11.2
pyparsing==3.0.8
pyrsistent==0.18.1
python-binance==1.0.15
python-dateutil==2.8.2
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
regex==2022.3.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
stack-data==0.2.0
telebot==0.0.4
telepot==12.7
tenacity==8.0.1
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
tzdata==2022.1
tzlocal==4.2
ujson==5.2.0
urllib3==1.26.9
wcwidth==0.2.5
websockets==9.1
yapf==0.32.0
yarl==1.7.2
